Add character reference decoder
Character references are used to specify any UTF-8 character in HTML by
codepoint.

If an invalid codepoint is given anywhere in a string, `CGI.unescape_html`
will raise a RangeError.

The method added here will examine a string for character references,
and when found replace with with the decoded UTF-8 character. If the
given codepoint is invalid, it will replace the character reference with
the empty string.

In Rets, this problem can arise when a parsed document contains a
character reference to an invalid codepoint that is HTML-encoded two
times. Nokogiri will decode it the first time, passing through the
now singly-encoded invalid character reference to
`CGI.html_unescape_html`.